Revert "Feat/1180 loading UI enhancements"
