+++ src/admin/components/elements/DatePicker/index.tsx
@@ -1,11 +1,11 @@
 import React, { Suspense, lazy } from 'react';
-import { ShimmerEffect } from '../ShimmerEffect';
+import Loading from '../Loading';
 import { Props } from './types';
 
 const DatePicker = lazy(() => import('./DatePicker'));
 
 const DatePickerField: React.FC<Props> = (props) => (
-  <Suspense fallback={<ShimmerEffect height={50} />}>
+  <Suspense fallback={<Loading />}>
     <DatePicker {...props} />
   </Suspense>
 );
