--- src/admin/components/elements/ShimmerEffect/index.scss
@@ -1,27 +1,0 @@
-.shimmer-effect {
-  position: relative;
-  overflow: hidden;
-  background-color: var(--theme-elevation-50);
-
-  &__shine {
-    position: absolute;
-    scale: 1.5;
-    width: 100%;
-    height: 100%;
-    transform: translateX(-100%);
-    animation: shimmer 1.75s infinite;
-    opacity: .75;
-    background: linear-gradient(100deg,
-        var(--theme-elevation-50) 0%,
-        var(--theme-elevation-50) 15%,
-        var(--theme-elevation-150) 50%,
-        var(--theme-elevation-50) 85%,
-        var(--theme-elevation-50) 100%)
-  }
-}
-
-@keyframes shimmer {
-  100% {
-    transform: translate3d(100%, 0, 0);
-  }
-}
