+++ src/admin/components/forms/Form/index.tsx
@@ -31,7 +31,6 @@ const baseClass = 'form';
 const Form: React.FC<Props> = (props) => {
   const {
     disabled,
-    disableNativeFormSubmission = true,
     onSubmit,
     method,
     action,
@@ -160,7 +159,7 @@ const Form: React.FC<Props> = (props) => {
       };
 
       onSubmit(fields, data);
-      if (disableNativeFormSubmission) return;
+      return;
     }
 
     const formData = contextRef.current.createFormData(overrides);
@@ -301,7 +300,6 @@ const Form: React.FC<Props> = (props) => {
     t,
     i18n,
     waitForAutocomplete,
-    disableNativeFormSubmission,
   ]);
 
   const getFields = useCallback(() => contextRef.current.fields, [contextRef]);
