+++ src/admin/components/forms/field-types/RichText/index.tsx
@@ -1,11 +1,11 @@
 import React, { Suspense, lazy } from 'react';
-import { ShimmerEffect } from '../../../elements/ShimmerEffect';
+import Loading from '../../../elements/Loading';
 import { Props } from './types';
 
 const RichText = lazy(() => import('./RichText'));
 
 const RichTextField: React.FC<Props> = (props) => (
-  <Suspense fallback={<ShimmerEffect height="35vh" />}>
+  <Suspense fallback={<Loading />}>
     <RichText {...props} />
   </Suspense>
 );
