+++ src/admin/components/views/Account/index.tsx
@@ -4,6 +4,7 @@ import { useTranslation } from 'react-i18next';
 import { useConfig } from '../../utilities/Config';
 import { useAuth } from '../../utilities/Auth';
 import { useStepNav } from '../../elements/StepNav';
+
 import usePayloadAPI from '../../../hooks/usePayloadAPI';
 import { useLocale } from '../../utilities/Locale';
 import DefaultAccount from './Default';
@@ -42,7 +43,7 @@ const AccountView: React.FC = () => {
 
   const { fields } = collection;
 
-  const [{ data, isLoading: isLoadingData }] = usePayloadAPI(
+  const [{ data }] = usePayloadAPI(
     `${serverURL}${api}/${slug}/${id}`,
     {
       initialParams: {
@@ -90,8 +91,6 @@ const AccountView: React.FC = () => {
     awaitInitialState();
   }, [dataToRender, fields, id, user, locale, preferencesKey, getPreference, t]);
 
-  const isLoading = !initialState || !docPermissions || isLoadingData;
-
   return (
     <RenderCustomComponent
       DefaultComponent={DefaultAccount}
@@ -104,7 +103,7 @@ const AccountView: React.FC = () => {
         hasSavePermission,
         initialState,
         apiURL,
-        isLoading,
+        isLoading: !initialState || !docPermissions,
         onSave,
       }}
     />
