+++ src/admin/components/views/Global/index.tsx
@@ -5,7 +5,9 @@ import { useConfig } from '../../utilities/Config';
 import { useAuth } from '../../utilities/Auth';
 import { useStepNav } from '../../elements/StepNav';
 import usePayloadAPI from '../../../hooks/usePayloadAPI';
+
 import { useLocale } from '../../utilities/Locale';
+
 import RenderCustomComponent from '../../utilities/RenderCustomComponent';
 import DefaultGlobal from './Default';
 import buildStateFromSchema from '../../forms/Form/buildStateFromSchema';
@@ -15,7 +17,7 @@ import { Fields } from '../../forms/Form/types';
 import { usePreferences } from '../../utilities/Preferences';
 
 const GlobalView: React.FC<IndexProps> = (props) => {
-  const { state: locationState } = useLocation<{ data?: Record<string, unknown> }>();
+  const { state: locationState } = useLocation<{data?: Record<string, unknown>}>();
   const locale = useLocale();
   const { setStepNav } = useStepNav();
   const { user } = useAuth();
@@ -55,7 +57,7 @@ const GlobalView: React.FC<IndexProps> = (props) => {
     setInitialState(state);
   }, [getVersions, fields, user, locale, t, getDocPermissions]);
 
-  const [{ data, isLoading: isLoadingData }] = usePayloadAPI(
+  const [{ data }] = usePayloadAPI(
     `${serverURL}${api}/globals/${slug}`,
     { initialParams: { 'fallback-locale': 'null', depth: 0, draft: 'true' } },
   );
@@ -80,14 +82,12 @@ const GlobalView: React.FC<IndexProps> = (props) => {
     awaitInitialState();
   }, [dataToRender, fields, user, locale, getPreference, preferencesKey, t]);
 
-  const isLoading = !initialState || !docPermissions || isLoadingData;
-
   return (
     <RenderCustomComponent
       DefaultComponent={DefaultGlobal}
       CustomComponent={CustomEdit}
       componentProps={{
-        isLoading,
+        isLoading: !initialState || !docPermissions,
         data: dataToRender,
         permissions: docPermissions,
         initialState,
