+++ src/admin/components/views/Version/index.tsx
@@ -1,17 +1,11 @@
 import React, { Suspense, lazy } from 'react';
-import { LoadingOverlayToggle } from '../../elements/Loading';
+import Loading from '../../elements/Loading';
 import { Props } from './types';
 
 const VersionView = lazy(() => import('./Version'));
 
 const Version: React.FC<Props> = (props) => (
-  <Suspense fallback={(
-    <LoadingOverlayToggle
-      show
-      name="version-suspense"
-    />
-  )}
-  >
+  <Suspense fallback={<Loading />}>
     <VersionView {...props} />
   </Suspense>
 );
