+++ src/admin/components/views/collections/Edit/index.tsx
@@ -59,7 +59,7 @@ const EditView: React.FC<IndexProps> = (props) => {
     }
   }, [admin, collection, isEditing, getVersions, user, id, t, locale, getDocPermissions]);
 
-  const [{ data, isLoading: isLoadingData, isError }] = usePayloadAPI(
+  const [{ data, isLoading: isLoadingDocument, isError }] = usePayloadAPI(
     (isEditing ? `${serverURL}${api}/${slug}/${id}` : null),
     { initialParams: { 'fallback-locale': 'null', depth: 0, draft: 'true' } },
   );
@@ -67,6 +67,9 @@ const EditView: React.FC<IndexProps> = (props) => {
   const dataToRender = (locationState as Record<string, unknown>)?.data || data;
 
   useEffect(() => {
+    if (isLoadingDocument) {
+      return;
+    }
     const awaitInitialState = async () => {
       setUpdatedAt(dataToRender?.updatedAt);
       const state = await buildStateFromSchema({ fieldSchema: fields, data: dataToRender, user, operation: isEditing ? 'update' : 'create', id, locale, t });
@@ -75,7 +78,7 @@ const EditView: React.FC<IndexProps> = (props) => {
     };
 
     awaitInitialState();
-  }, [dataToRender, fields, isEditing, id, user, locale, isLoadingData, preferencesKey, getPreference, t]);
+  }, [dataToRender, fields, isEditing, id, user, locale, isLoadingDocument, preferencesKey, getPreference, t]);
 
   useEffect(() => {
     if (redirect) {
@@ -92,7 +95,6 @@ const EditView: React.FC<IndexProps> = (props) => {
   const apiURL = `${serverURL}${api}/${slug}/${id}${collection.versions.drafts ? '?draft=true' : ''}`;
   const action = `${serverURL}${api}/${slug}${isEditing ? `/${id}` : ''}?locale=${locale}&depth=0&fallback-locale=null`;
   const hasSavePermission = (isEditing && docPermissions?.update?.permission) || (!isEditing && (docPermissions as CollectionPermission)?.create?.permission);
-  const isLoading = !initialState || !docPermissions || isLoadingData;
 
   return (
     <EditDepthContext.Provider value={1}>
@@ -101,7 +103,7 @@ const EditView: React.FC<IndexProps> = (props) => {
         CustomComponent={CustomEdit}
         componentProps={{
           id,
-          isLoading,
+          isLoading: !initialState || !docPermissions,
           data: dataToRender,
           collection,
           permissions: docPermissions,
