+++ src/admin/components/views/collections/List/Default.tsx
@@ -16,7 +16,6 @@ import PerPage from '../../../elements/PerPage';
 import { Gutter } from '../../../elements/Gutter';
 import { RelationshipProvider } from './RelationshipProvider';
 import { getTranslation } from '../../../../../utilities/getTranslation';
-import { StaggeredShimmers } from '../../../elements/ShimmerEffect';
 
 import './index.scss';
 
@@ -97,18 +96,6 @@ const DefaultList: React.FC<Props> = (props) => {
           handleSortChange={handleSortChange}
           handleWhereChange={handleWhereChange}
         />
-
-        {!data.docs && (
-          <StaggeredShimmers
-            className={[
-              `${baseClass}__shimmer`,
-              upload ? `${baseClass}__shimmer--uploads` : `${baseClass}__shimmer--rows`,
-            ].filter(Boolean).join(' ')}
-            count={6}
-            width={upload ? 'unset' : '100%'}
-          />
-        )}
-
         {(data.docs && data.docs.length > 0) && (
           <React.Fragment>
             {!upload && (
