+++ src/admin/index.tsx
@@ -18,7 +18,6 @@ import Routes from './components/Routes';
 import { StepNavProvider } from './components/elements/StepNav';
 import { ThemeProvider } from './components/utilities/Theme';
 import { I18n } from './components/utilities/I18n';
-import { LoadingOverlayProvider } from './components/utilities/LoadingOverlay';
 
 import './scss/app.scss';
 
@@ -46,11 +45,9 @@ const Index = () => (
                     <SearchParamsProvider>
                       <LocaleProvider>
                         <StepNavProvider>
-                          <LoadingOverlayProvider>
-                            <CustomProvider>
-                              <Routes />
-                            </CustomProvider>
-                          </LoadingOverlayProvider>
+                          <CustomProvider>
+                            <Routes />
+                          </CustomProvider>
                         </StepNavProvider>
                       </LocaleProvider>
                     </SearchParamsProvider>
