+++ src/translations/es.json
@@ -211,8 +211,6 @@
     "untitled": "Sin título",
     "updatedAt": "Fecha de modificado",
     "updatedSuccessfully": "Actualizado con éxito.",
-    "updating": "Actualizando",
-    "uploading": "Subiendo",
     "user": "usuario",
     "users": "usuarios",
     "welcome": "Bienvenido"
