+++ src/translations/pl.json
@@ -171,7 +171,7 @@
     "leaveAnyway": "Wyjdź mimo to",
     "leaveWithoutSaving": "Wyjdź bez zapisywania",
     "light": "Jasny",
-    "loading": "Ładowanie",
+    "loading": "Ładowanie...",
     "locales": "Lokalne",
     "moveDown": "Przesuń niżej",
     "moveUp": "Przesuń wyżej",
@@ -211,8 +211,6 @@
     "untitled": "Bez nazwy",
     "updatedAt": "Data edycji",
     "updatedSuccessfully": "Aktualizacja zakończona sukcesem.",
-    "updating": "Aktualizacja",
-    "uploading": "Wgrywanie",
     "user": "użytkownik",
     "users": "użytkownicy",
     "welcome": "Witaj"
