+++ src/translations/ru.json
@@ -181,7 +181,7 @@
     "leaveAnyway": "Все равно уйти",
     "leaveWithoutSaving": "Выход без сохранения",
     "light": "Светлая",
-    "loading": "Загрузка",
+    "loading": "Загрузка...",
     "locales": "Локали",
     "moveDown": "Сдвинуть вниз",
     "moveUp": "Сдвинуть вверх",
@@ -221,8 +221,6 @@
     "untitled": "Без названия",
     "updatedAt": "Дата правки",
     "updatedSuccessfully": "Успешно Обновлено.",
-    "updating": "Обновление",
-    "uploading": "Загрузка",
     "user": "пользователь",
     "users": "пользователи",
     "welcome": "Добро пожаловать"
