+++ src/translations/tr.json
@@ -138,6 +138,7 @@
     "none": "Hiç",
     "enterAValue": "Değer girin",
     "deletedSuccessfully": "Başarıyla silindi.",
+    "updatedSuccessfully": "Başarıyla güncellendi.",
     "aboutToDelete": "<1>{{title}}</1> {{label}} silinmek üzere. Silme işlemine devam etmek istiyor musunuz?",
     "addFilter": "Filtre ekle",
     "adminTheme": "Admin arayüzü",
@@ -183,7 +184,7 @@
     "leaveWithoutSaving": "Kaydetmeden ayrıl",
     "light": "Aydınlık",
     "locales": "Diller",
-    "loading": "Yükleniyor",
+    "loading": "Yükleniyor...",
     "moveDown": "Aşağı taşı",
     "moveUp": "Yukarı taşı",
     "newPassword": "Yeni parola",
@@ -220,15 +221,11 @@
     "unsavedChangesDuplicate": "Kaydedilmemiş değişiklikler var. Çoğaltma işlemine devam etmek istiyor musunuz?",
     "untitled": "Başlıksız",
     "updatedAt": "Güncellenme tarihi",
-    "updatedSuccessfully": "Başarıyla güncellendi.",
-    "updating": "Güncelleniyor",
-    "uploading": "Yükleniyor",
     "user": "kullanıcı",
     "users": "kullanıcı",
     "welcome": "Hoşgeldiniz"
   },
   "upload": {
-    "dragAndDrop": "Bir dosya sürükleyip bırakabilirsiniz",
     "dragAndDropHere": "veya buraya bir dosya sürükleyip bırakabilirsiniz",
     "fileName": "Dosya adı",
     "fileSize": "Dosya boyutu",
