+++ src/translations/translation-schema.json
@@ -797,12 +797,6 @@
         "updatedSuccessfully": {
           "type": "string"
         },
-        "updating": {
-          "type": "string"
-        },
-        "uploading": {
-          "type": "string"
-        },
         "user": {
           "type": "string"
         },
